requirements.txt
tkinter
pygame
pillow
opencv-python
argparse
natsort
numpy
tqdm